requirements.txt
Pillow
jupyter